Move ReadOnlyAdminMixin to the core app
