+++ hipogram/core/mixins.py
@@ -1,0 +1,11 @@
+
+class ReadOnlyAdminMixin:
+
+    def has_add_permission(self, *args, **kwargs):
+        return False
+
+    def has_change_permission(self, *args, **kwargs):
+        return False
+
+    def has_delete_permission(self, *args, **kwargs):
+        return False
